#5936 - Explorer reports involving document-level features remain blank
- Fix document name and string feature extractors to work on any kind of feature structures
