+++ inception/inception-pivot-api/src/main/java/de/tudarmstadt/ukp/inception/pivot/api/extractor/FeatureStructureExtractor_ImplBase.java
@@ -1,0 +1,73 @@
+/*
+ * Licensed to the Technische Universität Darmstadt under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The Technische Universität Darmstadt 
+ * licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.
+ *  
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.tudarmstadt.ukp.inception.pivot.api.extractor;
+
+import static java.util.Optional.empty;
+
+import java.io.Serializable;
+import java.util.Optional;
+
+import org.apache.uima.cas.FeatureStructure;
+
+import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
+
+public abstract class FeatureStructureExtractor_ImplBase<T extends FeatureStructure, R extends Serializable>
+    implements Extractor<ContextualizedFS<T>, R>
+{
+    private final AnnotationLayer layer;
+
+    public FeatureStructureExtractor_ImplBase(AnnotationLayer aLayer)
+    {
+        layer = aLayer;
+    }
+
+    public AnnotationLayer getLayer()
+    {
+        return layer;
+    }
+
+    @Override
+    public Optional<String> getTriggerType()
+    {
+        if (layer == null) {
+            return empty();
+        }
+
+        return Optional.of(layer.getName());
+    }
+
+    @Override
+    public boolean isWeak()
+    {
+        return layer == null;
+    }
+
+    @Override
+    public boolean accepts(Object aSource)
+    {
+        if (aSource instanceof ContextualizedFS ann) {
+            if (layer == null) {
+                return true;
+            }
+
+            return layer.getName().equals(ann.fs().getType().getName());
+        }
+
+        return false;
+    }
+}
